% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendData.R
\name{appendFluVaxEfficacyData}
\alias{appendFluVaxEfficacyData}
\title{appendFluVaxEfficacyData: internal function for adding model$summary.random to db$observedData from fluVaxEfficacyModel fit}
\usage{
appendFluVaxEfficacyData(model, modelDefinition)
}
\arguments{
\item{model}{inla model object}

\item{db}{object from dbViewer with observedData tibble and query}
}
\value{
db with added modeledData tibble
}
\description{
appendFluVaxEfficacyData: internal function for adding model$summary.random to db$observedData from fluVaxEfficacyModel fit
}
