% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendILIDataFc.R
\name{appendILIDataFc}
\alias{appendILIDataFc}
\title{appendILIDataFc: function for adding ILI Data and short fc to db$observedData}
\usage{
appendILIDataFc(db, currentWeek)
}
\arguments{
\item{db}{object from dbViewer with observedData tibble and query}

\item{currentWeek}{current week}
}
\value{
db with added ILI Data + fc tibble
}
\description{
appendILIDataFc: function for adding ILI Data and short fc to db$observedData
}
