% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentFieldModel.R
\name{appendLatentFieldData}
\alias{appendLatentFieldData}
\title{appendLatentFieldData: internal function for adding model$summary.random to db$observedData from latentFieldModel fit}
\usage{
appendLatentFieldData(model, modelDefinition)
}
\arguments{
\item{model}{inla model object}

\item{db}{object from dbViewer with observedData tibble and query}
}
\value{
db with added modeledData tibble
}
\description{
appendLatentFieldData: internal function for adding model$summary.random to db$observedData from latentFieldModel fit
}
