% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTrainR.R
\name{appendModelData}
\alias{appendModelData}
\title{appendModelData: internal function for adding model$summary.fitted.values to db}
\usage{
appendModelData(model, db)
}
\arguments{
\item{model}{inla model object}

\item{db}{object from dbViewer with observedData tibble and query}
}
\value{
db with added modeledData tibble
}
\description{
appendModelData: internal function for adding model$summary.fitted.values to db
}
