% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentFieldModel.R, R/smoothModel.R
\name{appendSmoothData}
\alias{appendSmoothData}
\title{appendSmoothData: internal function for adding model$summary.fitted.values to db$observedData from smoothModel fit}
\usage{
appendSmoothData(model, db, family = "poisson")

appendSmoothData(model, db, family = "poisson")
}
\arguments{
\item{model}{inla model object}

\item{db}{object from dbViewer with observedData tibble and query}

\item{model}{inla model object}

\item{db}{object from dbViewer with observedData tibble and query}
}
\value{
db with added modeledData tibble

db with added modeledData tibble
}
\description{
appendSmoothData: internal function for adding model$summary.fitted.values to db$observedData from smoothModel fit

appendSmoothData: internal function for adding model$summary.fitted.values to db$observedData from smoothModel fit
}
