% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ILI.R
\name{calcfc}
\alias{calcfc}
\title{calcfc: internal function for calculating the ILI fit with no regressors}
\usage{
calcfc(dat, num_prior_data = 52, currentWeek = currentWeek)
}
\arguments{
\item{dat}{dataframe with week, and ILI as columns}

\item{num_prior_data}{number of previous data points (weeks) to use to fit the ARIMA model}

\item{num_future_data}{number of weeks forward to make ILI estimates}
}
\value{
ILI_estimate - dataframe with point estimate and 95% CI upper and lower for num_future_weeks
}
\description{
takes logit of ILI data for modeling, inverse logit is done before returning ILI estimates
uses num_prior_weeks - number of points to fit for each data point in the 2018-2019 season
forward estimates for num_future_weeks - number of weeks to make future ILI estimates
returns ILI forward estimates for num_future_weeks- point estimate and the 95\% CI upper and lower
}
