% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTrainR.R
\name{constructAdjacencyNetwork}
\alias{constructAdjacencyNetwork}
\title{constructAdjacencyNetwork:  function for constructing adjacency graph for INLA GMRF from shapefile}
\usage{
constructAdjacencyNetwork(shp)
}
\arguments{
\item{shp}{sf object with polygon geometry}
}
\value{
neighborGraph in INLA format
}
\description{
constructAdjacencyNetwork:  function for constructing adjacency graph for INLA GMRF from shapefile
}
\examples{
return neighbor census tract adjacency network for King County, WA
   neighborGraph <- constructAdjacencyNetwork( shp = dbViewR::masterSpatialDB() )

}
