% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluVaxEfficacyModel.R
\name{fluVaxEfficacyModel}
\alias{fluVaxEfficacyModel}
\title{fluVaxEfficacyModel: function to define latent field model tailored to test-negative study
for flu vax efficacy from dbViewR object}
\usage{
fluVaxEfficacyModel(db, shp = NULL, neighborGraph = NULL)
}
\arguments{
\item{db}{dbViewR object with valid column names for INLA model. required fields: "pathogen" and "flu_shot"}

\item{shp}{sf object with residence_census_tract shapes (all higher levels assume iid and not local smoothing)}

\item{neighborGraph}{non-standard neighbor graph (default = NULL)}

\item{family}{non-standard family override (default = NULL).}
}
\value{
modelDefinition object for modelTrainR, list with fields
    type = vaccine_efficacy
    family : as input
    formula : model definition for INLA
    inputData : inla-prepared db$observed data
    neighborGraph : as input or derived from shp during formula construction
}
\description{
fluVaxEfficacyModel: function to define latent field model tailored to test-negative study
for flu vax efficacy from dbViewR object
}
\examples{
return h1n1pdm incidence model by time and location
   modelDefinition <- smoothModel(db = dbViewR::selectFromDB(), shp = dbViewR::masterSpatialDB())

}
