% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ILI.R
\name{forecast_ILI}
\alias{forecast_ILI}
\title{forecast_ILI: function to extend CDC ILI data to desired week via ARIMA forecast}
\usage{
forecast_ILI(num_prior_data = 52, currentWeek = NULL)
}
\arguments{
\item{num_prior_data}{number of previous data points to use to fit the ARIMA model}

\item{num_future_data}{number of weeks forward from the latest observation in the CDC database to make ILI estimates}
}
\value{
ILI_estimate - dataframe with point estimate and 95% CI upper and lower for num_future_data weeks
}
\description{
forecast_ILI: function to extend CDC ILI data to desired week via ARIMA forecast
}
\examples{
return ILI estimate 2 weeks ahead
   ILI_forecast <- forecast_ILI(num_prior_data=52, num_future_weeks = 2)
   
   
}
