% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTrainR.R
\name{modelTrainR}
\alias{modelTrainR}
\title{modelTrainR: function for training INLA models from modelDefintion objects}
\usage{
modelTrainR(modelDefinition)
}
\arguments{
\item{modelDefinition}{modelDefinition object from one of smoothModel.R, latentFieldModel.R, or effectModel.R}
}
\value{
observedData table that has been prepared for defineModels.R
}
\description{
modelTrainR: function for training INLA models from modelDefintion objects
}
\examples{
return h1n1pdm incidence model by time and location
   db <- modelTrainR(db = dbViewR::selectFromDB(), shp = dbViewR::masterSpatialDB())

}
