% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_ILI.R
\name{plotforecasts}
\alias{plotforecasts}
\title{plotforecasts: internal function for saving plot of ILI and ILI forecast with 95\%CI}
\usage{
plotforecasts(dat, fitted_param, ggtitlestr = NULL, outpath = NULL)
}
\arguments{
\item{dat}{ILI data to plot}

\item{fitted_param}{ILI}

\item{ggtitlestr}{title of plot}

\item{outpath}{where to save}
}
\description{
plotforecasts: internal function for saving plot of ILI and ILI forecast with 95\%CI
}
\examples{
plot ILI data and forecast
  plotforecasts(dat_WAILI, "ILI", "ARIMA forecast ILI", "ARIMA_fc_ILI")
}
