% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTrainR.R
\name{saveModel}
\alias{saveModel}
\title{saveModel: function to save models and register them in modelDB.csv}
\usage{
saveModel(model,
  cloudDir = "C:/Users/mfamulare/Dropbox (IDM)/SeattleFlu-incidenceMapR/models")
}
\arguments{
\item{model}{INLA object}

\item{db}{dbViewR object}
}
\description{
saveModel: function to save models and register them in modelDB.csv
}
