% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothModel.R
\name{smoothModel}
\alias{smoothModel}
\title{smoothModel: function to define variable smoothing models from dbViewR object}
\usage{
smoothModel(db, shp, family = NULL, neighborGraph = NULL)
}
\arguments{
\item{db}{dbViewR object with valid column names for INLA model.
Smoothing only makes sense by age, location, and time.  Factor variables cannot be smoothed!}

\item{shp}{sf object with residence_census_tract shapes (all higher levels assume iid and not local smoothing)}

\item{family}{non-standard family override (default = NULL).}

\item{neighborGraph}{non-standard neighbor graph (default = NULL)}
}
\value{
modelDefinition object for modelTrainR, list with fields
    type = smooth
    family : as input
    formula : model definition for INLA
    inputData : inla-prepared db$observed data
    neighborGraph : as input or derived from shp during formula construction
}
\description{
smoothModel: function to define variable smoothing models from dbViewR object
}
\examples{
return h1n1pdm incidence model by time and location
   modelDefinition <- smoothModel(db = dbViewR::selectFromDB(), shp = dbViewR::masterSpatialDB())

}
