% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTrainR.R
\name{modelTrainR}
\alias{modelTrainR}
\title{modelTrainR: function for training binomial (logistic) regression models on data from dbViewR}
\usage{
modelTrainR(family = "binomial", db = dbViewR::selectFromDB(),
  shp = dbViewR::masterSpatialDB(), formula = NULL, inputData = NULL,
  neighborGraph = NULL)
}
\arguments{
\item{family}{regression model family (binomial (default) or poisson)}

\item{db}{dbViewR object with valid column names for INLA model}

\item{shp}{sf object with GEOID shapes (all higher levels assume iid and not local smoothing)}
}
\value{
observedData table that has been prepared for defineModels.R
}
\description{
modelTrainR: function for training binomial (logistic) regression models on data from dbViewR
}
\examples{
return h1n1pdm incidence model by time and location
   db <- modelTrainR(db = dbViewR::selectFromDB(), shp = dbViewR::masterSpatialDB())

}
